Implement A-Assertions
Current code does not document important assumptions that tasks should
not be null and task size should be more than or equal to 0.

Implementing assertions will define these assumptions about the program
state so that the runtime can verify them.

Let's,
* create the assertion in the TaskList class that task size should be
more than or equal to 0
* create the assertion in the Ui class and TaskList class that tasks
should not be null